Added some additional constructors.  Allows each transition to specify whether it should be treated as an XTraceProcess or not - ie. whether to add an edge from the transition event to the previous transition event.
If the instigation of a transition will always end in another transition
(ie. Running -> Failed, or Running -> Success) then we should log it as
a process.  

If the instigation of a transition is just administrative (ie.
Initialization, or Status Update) then we do not log as a process.

This commit will be followed up by modifications of each of the Hadoop
state machines on a case by case basis, to determine whether transitions
are processes or not.